requirements.txt
flask==3.0.0
pytest==4.0.1
Jinja2==3.1.1
bitlyshortener
validators
python-dotenv